Fixes pie chart so it uses correct data and renames. Moved header and nav components into shared components folder
